refactor(explain): extract the plan model and parsers into tabularis-explain
The EXPLAIN plan model and every parser lived next to the code that runs
EXPLAIN statements: `ExplainPlan` sat in `models.rs` alongside connection
types, the Postgres parsers were mixed into the Tauri file-import commands,
and ~900 lines of pure MySQL/MariaDB JSON and tree-format parsing were
interleaved with `sqlx` execution in the driver.

Move all of it to a standalone `tabularis-explain` crate whose only
dependencies are `serde` and `serde_json`. It converts already-produced
EXPLAIN output into a plan tree and knows nothing about connections,
statement building, async runtimes or the filesystem — so the same parsers
can also be compiled to WASM for a browser-only plan visualiser.

What stayed behind, deliberately:

- `explain_import.rs` keeps the Tauri commands, the standalone window and
  the file read — host concerns, not parsing.
- `drivers/*/explain.rs` keep statement building and execution.
- MySQL's tabular `EXPLAIN` parser stays in the driver: it only ever
  arrives as decoded `sqlx` rows, never as a serialisable payload.

`models.rs` re-exports `ExplainNode` / `ExplainPlan`, so the twelve modules
importing them are untouched. Parser tests move with their parsers: 31 of
them now run against the crate's public API with no driver or host in
scope.

Kept out of a Cargo workspace on purpose — a workspace root would move the
build directory away from `src-tauri/target` (hardcoded in release.yml) and
silently override `src-tauri`'s `[profile.release]`.
