+++ crates/tabularis-explain/src/postgres/mod.rs
@@ -1,0 +1,7 @@
+//! Postgres plan parsers.
+
+mod json;
+mod text;
+
+pub use json::parse_postgres_json;
+pub use text::parse_postgres_text;
