+++ src-tauri/Cargo.toml
@@ -23,6 +23,9 @@ strip = true
 tauri-build = { version = "2.5.3", features = [] }
 
 [dependencies]
+# Plan model + parsers, kept in a standalone crate so they stay free of any
+# query-execution or host concern. See crates/tabularis-explain/README.md.
+tabularis-explain = { path = "../crates/tabularis-explain" }
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 log = "0.4"
